Use uint8

+++ Inc/BSMath/SIMD.h
@@ -125,24 +125,24 @@ namespace BSMath::SIMD
     template <Swizzle X, Swizzle Y, Swizzle Z, Swizzle W>
     [[nodiscard]] NO_ODR VectorRegister<int> VECTOR_CALL VectorSwizzle(VectorRegister<int> vec) noexcept
     {
-        constexpr static auto mask = static_cast<uint8>(_MM_SHUFFLE(static_cast<uint32>(W),
-            static_cast<uint32>(Z), static_cast<uint32>(Y), static_cast<uint32>(X)));
+        constexpr static auto mask = _MM_SHUFFLE(static_cast<uint32>(W),
+            static_cast<uint32>(Z), static_cast<uint32>(Y), static_cast<uint32>(X));
         return _mm_shuffle_epi32(vec, mask);
     }
 
     template <Swizzle X, Swizzle Y, Swizzle Z, Swizzle W>
     [[nodiscard]] NO_ODR VectorRegister<float> VECTOR_CALL VectorShuffle(VectorRegister<float> lhs, VectorRegister<float> rhs) noexcept
     {
-        constexpr static auto mask = static_cast<uint8>(_MM_SHUFFLE(static_cast<uint32>(W),
-            static_cast<uint32>(Z), static_cast<uint32>(Y), static_cast<uint32>(X)));
+        constexpr static auto mask = _MM_SHUFFLE(static_cast<uint8>(W),
+            static_cast<uint8>(Z), static_cast<uint8>(Y), static_cast<uint8>(X));
         return _mm_shuffle_ps(lhs, rhs, mask);
     }
 
     template <Swizzle X, Swizzle Y, Swizzle Z, Swizzle W>
     [[nodiscard]] NO_ODR VectorRegister<int> VECTOR_CALL VectorShuffle(VectorRegister<int> lhs, VectorRegister<int> rhs) noexcept
     {
-        constexpr static auto mask = _MM_SHUFFLE(static_cast<uint32>(W),
-            static_cast<uint32>(Z), static_cast<uint32>(Y), static_cast<uint32>(X));
+        constexpr static auto mask = _MM_SHUFFLE(static_cast<uint8>(W),
+            static_cast<uint8>(Z), static_cast<uint8>(Y), static_cast<uint8>(X));
         return _mm_castps_si128(_mm_shuffle_ps(_mm_castsi128_ps(lhs), _mm_castsi128_ps(rhs), mask));
     }
